+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### 6.8.3 (9/13/2018 - )
 * http: removed request.pathParam/request.queryParam, for type safety, we use api interface/impl, so only keep low level interface to keep flexibility 
 * http: path param now only supports string/integer/number/enum
+* kafka: bulk message handler logs all message in trace for troubleshooting 
 
 ### 6.8.2 (9/10/2018 - 9/13/2018)
 * http: change framework bad request error code to "INVALID_HTTP_REQUEST"